process-data/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
autosub==0.3.12
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.3
colorama==0.4.6
colorlog==6.8.2
contourpy==1.1.1
coverage==5.5
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
dlib @ file:///E:/Do%20An/process-data/dlib-19.22.99-cp38-cp38-win_amd64.whl#sha256=b02321413b8765159ae552967ab9781a981324b2a389924bdb8c4ccab227f160
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.2
eyed3==0.9.7
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
groq==0.9.0
h11==0.14.0
hmmlearn==0.3.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.3
HyperPyYAML==1.2.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
imbalanced-learn==0.12.3
imblearn==0.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.6
llvmlite==0.41.1
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==6.0.5
networkx==3.1
noisereduce==3.0.2
numba==0.58.1
numpy==1.24.4
omegaconf==2.3.0
openai-whisper==20231117
opencv-python==4.9.0.80
optuna==3.6.1
packaging==24.0
pandas==2.0.3
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pooch==1.8.1
primePy==1.3
proglog==0.1.10
progressbar==2.5
proto-plus==1.23.0
protobuf==4.25.3
pyannote.audio==3.3.0
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyAudioAnalysis==0.3.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
pysrt==1.1.2
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytorch-lightning==2.3.3
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
reportlab==4.2.2
requests==2.32.2
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.3.2
scipy==1.10.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
speechbrain==1.0.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.3.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.1
torchmetrics==1.4.0.post0
tqdm==4.66.4
transformers==4.41.2
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.0
